--- webapp/components/survey/NodeDefDisplayTypeSelector/index.js
@@ -1,1 +1,0 @@
-export { default } from './NodeDefDisplayTypeSelector'
